nginx: Create map to store requests and their trace contexts
Nginx requests share variables with their subrequests so we should not
create contexts without first checking if it is a subrequest. As the
variable is shared between the requests, we can instead create a map
that will record which request matches to which context so that we can
retrieve the right tracing context when FinishNgxSpan is called.

Signed-off-by: Remy Suen <remy.suen@gmail.com>